backend/requirements.txt
Flask 
Flask-cors 
torch 
torchvision 
transformers 
Pillow 
python-dotenv
PyJWT 
bcrypt